+++ lib/appium_lib_core/common/base/driver.rb
@@ -960,7 +960,6 @@ module Appium
           @bridge.session_capabilities
         end
 
-
         # Returns available sessions on the Appium server
         #
         # @return [[Hash]]