+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Read `release_notes.md` for commit level details.
 ## [Unreleased]
 
 ### Enhancements
+- Add `Logs#event` to post a custom log. `@driver.logs.event vendor: 'appium', event: 'funEvent'`
+    - It can get via `@driver.session_capabilities['events']`
 
 ### Bug fixes